feat: add app.ts
