+++ src/shared/infra/http/app.ts
@@ -1,0 +1,42 @@
+import express, { Request, Response, NextFunction } from 'express';
+import 'express-async-errors';
+import cors from 'cors';
+import { errors } from 'celebrate';
+import { pagination } from 'typeorm-pagination';
+import routes from './routes';
+import AppError from '@shared/errors/AppError';
+import '@shared/infra/typeorm';
+import '@shared/container';
+import uploadConfig from '@config/upload';
+import rateLimiter from './middlewares/rateLimiter';
+
+const app = express();
+
+app.use(cors());
+app.use(express.json());
+app.use(rateLimiter);
+app.use(pagination);
+
+app.use('/files', express.static(uploadConfig.directory));
+app.use(routes);
+
+app.use(errors());
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+app.use((error: Error, req: Request, res: Response, next: NextFunction) => {
+  if (error instanceof AppError) {
+    return res.status(error.statusCode).json({
+      status: 'error',
+      message: error.message,
+    });
+  }
+
+  console.log(error);
+
+  return res.status(500).json({
+    status: 'error',
+    message: 'Internal Server Error',
+  });
+});
+
+export { app };
