+++ src/shared/infra/http/server.ts
@@ -1,47 +1,10 @@
 import 'reflect-metadata';
 import 'dotenv/config';
-import express, { Request, Response, NextFunction } from 'express';
-import 'express-async-errors';
-import cors from 'cors';
-import { errors } from 'celebrate';
-import { pagination } from 'typeorm-pagination';
-import routes from './routes';
-import AppError from '@shared/errors/AppError';
-import '@shared/infra/typeorm';
-import '@shared/container';
-import uploadConfig from '@config/upload';
-import rateLimiter from './middlewares/rateLimiter';
+import { app } from './app';
+import { dataSource } from '../typeorm';
 
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-app.use(rateLimiter);
-app.use(pagination);
-
-app.use('/files', express.static(uploadConfig.directory));
-app.use(routes);
-
-app.use(errors());
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-app.use((error: Error, req: Request, res: Response, next: NextFunction) => {
-  if (error instanceof AppError) {
-    return res.status(error.statusCode).json({
-      status: 'error',
-      message: error.message,
-    });
-  }
-
-  console.log(error);
-
-  return res.status(500).json({
-    status: 'error',
-    message: 'Internal Server Error',
+dataSource.initialize().then(() => {
+  app.listen(process.env.PORT || 3333, () => {
+    console.log(`Server started on port ${process.env.PORT || 3333}! 🏆`);
   });
-});
-
-app.listen(3333, () => {
-  // eslint-disable-next-line no-console
-  console.log('Server started on port 3333! 🏆');
 });
